Separate cli from library
Split code into main library file and commmand line interface file.
